transform view functions into classes
