--- src/app/views/renderNotFound.ts
@@ -1,3 +1,0 @@
-export default function renderNotFound(el: HTMLElement): void {
-  el.innerHTML = '<p class=\'text-red\'>Error 404: la página solicitada no fue encontrada.</p>';
-}
